requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Frontend
streamlit==1.31.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# LLM Integration (using Groq - free and fast)
groq==0.4.2

# Web Scraping / HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Selenium for testing
selenium==4.17.2

# Utilities
requests==2.31.0
pydantic==2.6.0
python-dotenv==1.0.1